requirements.txt
# Grundläggande
python-dotenv==1.0.1
tiktoken==0.5.2

# Langchain + community
langchain==0.1.14
langchain-community==0.0.30
langchain-openai==0.0.5
langchain_community

# Dokumentladdare & markdown-stöd
unstructured[md]==0.10.28

# Embeddinglagring
chromadb==0.5.0
onnxruntime==1.21.0

# FastAPI + server
fastapi
uvicorn
httpx

# OpenAI (låt pip installera rätt version)
openai
